requirements.txt
flask~=3.1.1
pandas~=2.2.3
openpyxl
requests~=2.32.4
sqlalchemy~=2.0.41
python-dotenv~=1.1.0
python-telegram-bot~=22.1
flask-cors~=6.0.1